+++ src/Parsers/DetourParser.php
@@ -27,7 +27,7 @@ class DetourParser
             $this->geometryParser->parse($object->geometry),
             ($object->properties->name ?? '') ?: sprintf('Omleiding %d', $index + 1),
             Direction::from($object->properties->direction),
-            property_exists($object->properties, 'transportMode') ? TransportMode::from($object->properties->transportMode) : null,
+            !empty($object->properties->transportMode) ? TransportMode::from($object->properties->transportMode) : null,
             array_map([VehicleType::class, 'from'], $object->properties->vehicleTypes),
             array_map([BoatType::class, 'from'], $object->properties->boatTypes)
         );